requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-ollama
langchain_community
beautifulsoup4
chromadb
httptools
pip-chill
python-dotenv
#uvloop==0.19.0
watchfiles
websockets
# other required dependenciesl